backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# MongoDB async driver
motor==3.3.2
pymongo==4.6.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data validation
pydantic==2.9.0
pydantic-settings==2.1.0

# HTTP client for testing
httpx==0.25.2

# Docker SDK for container management
docker==6.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development tools
black==23.11.0
ruff==0.1.6
mypy==1.7.1
pre-commit==3.5.0

# Logging and monitoring
structlog==23.2.0
